==> 0.InitializeData.ipynb <==
import os
from os import path
import zipfile

# Load initial data

data_archive = './titanic.zip'
data_folder = './source_data'
if path.exists(data_archive):
    os.remove(data_archive)
os.system('kaggle competitions download -c titanic')
with zipfile.ZipFile(data_archive, 'r') as zip_ref:
    zip_ref.extractall(data_folder)
if path.exists(data_archive):
    os.remove(data_archive)
