==> 1.PreprocessData.ipynb <==
# Initial data analysis

First a bit of description:

<table>
<tbody>
<tr><th><b>Variable</b></th><th><b>Definition</b></th><th><b>Key</b></th></tr>
<tr>
<td>survival</td>
<td>Survival</td>
<td>0 = No, 1 = Yes</td>
</tr>
<tr>
<td>pclass</td>
<td>Ticket class</td>
<td>1 = 1st, 2 = 2nd, 3 = 3rd</td>
</tr>
<tr>
<td>sex</td>
<td>Sex</td>
<td></td>
</tr>
<tr>
<td>Age</td>
<td>Age in years</td>
<td></td>
</tr>
<tr>
<td>sibsp</td>
<td># of siblings / spouses aboard the Titanic</td>
<td></td>
</tr>
<tr>
<td>parch</td>
<td># of parents / children aboard the Titanic</td>
<td></td>
</tr>
<tr>
<td>ticket</td>
<td>Ticket number</td>
<td></td>
</tr>
<tr>
<td>fare</td>
<td>Passenger fare</td>
<td></td>
</tr>
<tr>
<td>cabin</td>
<td>Cabin number</td>
<td></td>
</tr>
<tr>
<td>embarked</td>
<td>Port of Embarkation</td>
<td>C = Cherbourg, Q = Queenstown, S = Southampton</td>
</tr>
</tbody>
</table>

import numpy as np
import pandas as pd
import os
from os import path

data_folder = './source_data'
raw_train = pd.read_csv(path.join(data_folder, 'train.csv'), index_col='PassengerId').rename(str.lower, axis='columns')
raw_test = pd.read_csv(path.join(data_folder, 'test.csv'), index_col='PassengerId').rename(str.lower, axis='columns')
raw_train.index.names = raw_test.index.names = ['passengerid']

raw_train.head(20)

raw_test.head(20)

First let's realign the columns

survival = raw_train['survived']
train_test = pd.concat([raw_train, raw_test])
train_test = train_test[['pclass', 'name', 'sex', 'age', 'sibsp', 'parch', 'ticket', 'fare', 'cabin', 'embarked']]
train_test

Than let's make the data more convenient to deal with

def preprocess_df(df: pd.DataFrame) -> pd.DataFrame :
    pclass_ohe = pd.get_dummies(df['pclass'])
    pclass_ohe.columns = ['1cl', '2cl', '3cl']
    df = df.join(pclass_ohe)
    df['title'] = df['name'].str.extract(r'([A-z]+\.)')
    df['title'] = df['title'].str.lower()
    title_ohe = pd.get_dummies(df['title'])
    df = df.join(title_ohe)
    sex_ohe = pd.get_dummies(df['sex'])
    df = df.join(sex_ohe)
    df['deck'] = df['cabin'].str.extract(r'(^[A-z]+)')
    df['room'] = df['cabin'].str.extract(r'([0-9]+$)')
    df['ticket_srs'] = df['ticket'].str.replace(r'([0-9]+$)', '', regex=True)
    df['ticket_srs'] = df['ticket_srs'].str.strip()
    df['ticket_num'] = df['ticket'].str.extract(r'([0-9]+$)')
    deck_ohe = pd.get_dummies(df['deck']).rename(str.lower, axis='columns')
    df = df.join(deck_ohe)
    port_ohe = pd.get_dummies(df['embarked'])
    port_ohe.columns = ['cherbourg', 'queenstown', 'southampton']
    df = df.join(port_ohe)
    
    return df

prepr_train_test = preprocess_df(train_test)

prepr_train_test

Now let's make it clear about nulls and try to avoid them

prepr_train_test.shape

prepr_train_test.isna().sum()

prepr_train_test['age'] = prepr_train_test['age'].fillna(prepr_train_test.groupby('title')['age'].transform('mean'))
prepr_train_test['fare'] = prepr_train_test['fare'].fillna(prepr_train_test['fare'].mean())
prepr_train_test['embarked'] = prepr_train_test['embarked'].fillna('U')
prepr_train_test['ticket_num'] = prepr_train_test['ticket_num'].fillna(0)
prepr_train_test['cabin'] = prepr_train_test['cabin'].fillna('U0')
prepr_train_test['deck'] = prepr_train_test['deck'].fillna('U')
prepr_train_test['room'] = prepr_train_test['room'].fillna('0')

prepr_train_test.isna().sum()

Finally let's separate the data back to the train and test sets and save them.

prepr_train = prepr_train_test.loc[raw_train.index.values]
prepr_train['survived'] = survival
prepr_train

prepr_test = prepr_train_test.loc[raw_test.index.values]
prepr_test

prepr_train.to_csv(path.join(data_folder, 'prepr_train.csv'), index_label='passengerid')
prepr_test.to_csv(path.join(data_folder, 'prepr_test.csv'), index_label='passengerid')
