==> 2.BaselineModels.ipynb <==
# Applying baseline models to the prepared data

import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
import xgboost
import lightgbm
import catboost

from os import path
import os

Load the data...

data_folder = './source_data'
data = pd.read_csv(path.join(data_folder, 'prepr_train.csv'), index_col='passengerid').rename(str.lower, axis='columns')
data

num_model_features = ['age', 'sibsp', 'parch', 'fare', '1cl', '2cl', '3cl', 'capt.', 'col.',\
                      'countess.', 'don.', 'dona.', 'dr.', 'jonkheer.', 'lady.', 'major.', 'master.',\
                      'miss.', 'mlle.', 'mme.', 'mr.', 'mrs.', 'ms.', 'rev.', 'sir.', 'female', 'male',\
                      'room', 'ticket_num', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 't']
tree_model_features = ['pclass', 'sex', 'age', 'sibsp', 'parch', 'fare', 'embarked',\
                       'title', 'room', 'deck', 'ticket_srs', 'ticket_num']
col_mapper = {}

Now let's try the particular algorithms! The details could be found here:

https://scikit-learn.org/stable/modules/generated/sklearn.svm.SVC.html  
https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html  
https://scikit-learn.org/stable/modules/generated/sklearn.neighbors.KNeighborsClassifier.html  
https://scikit-learn.org/stable/modules/naive_bayes.html  
https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.RandomForestClassifier.html  
https://xgboost.readthedocs.io/en/latest/get_started.html  
https://lightgbm.readthedocs.io/en/latest/pythonapi/lightgbm.LGBMClassifier.html  
https://catboost.ai/en/docs/concepts/python-usages-examples  

from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import accuracy_score
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB, MultinomialNB, ComplementNB, BernoulliNB, CategoricalNB
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier
from sklearn.tree import DecisionTreeClassifier
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier

def check_model_kfold(model, data, mapper):
    kf = StratifiedKFold(n_splits=5)
    if model.__class__.__name__ in mapper.keys():
        X, y = data[mapper[model.__class__.__name__]], data['survived']
    else:
        X, y = data[num_model_features], data['survived']
    acc_score = []
    for train, test in kf.split(X, y):
        X_train, y_train, X_test, y_test = X.iloc[train], y.iloc[train], X.iloc[test], y.iloc[test]
        model.fit(X_train, y_train)
        y_pred = model.predict(X_test)
        acc = accuracy_score(y_pred, y_test)
        acc_score += [acc]
    return acc_score

def check_model_100(model, data, mapper):
    if model.__class__.__name__ in mapper.keys():
        X, y = data[mapper[model.__class__.__name__]], data['survived']
    else:
        X, y = data[num_model_features], data['survived']
    acc_score = []
    model.fit(X, y)
    y_pred = model.predict(X)
    acc = accuracy_score(y_pred, y)
    acc_score += [acc]
    return acc_score

models = [
    LogisticRegression(penalty='l1', solver='saga', random_state=0),\
    SVC(kernel='rbf'),\
    KNeighborsClassifier(),\
    GaussianNB(), MultinomialNB(), ComplementNB(), BernoulliNB(),\
    RandomForestClassifier(random_state=0), ExtraTreesClassifier(random_state=0), DecisionTreeClassifier(random_state=0),\
    XGBClassifier(random_state=0, verbosity=0), 
    LGBMClassifier(silent=True, verbose=-100), 
    CatBoostClassifier(random_state=0, silent=True)
]
for model in models:
    acc_score = check_model_kfold(model, data, col_mapper)
    print(f'\n{model.__class__.__name__}:')
    print(f'mean_score = {np.mean(acc_score)}; acc_score = {acc_score}')

Now let's make forecasts.

test = pd.read_csv(path.join(data_folder, 'prepr_test.csv'), index_col='passengerid').rename(str.lower, axis='columns')
test

def apply_model(model, data, mapper):
    if model.__class__.__name__ in mapper.keys():
        X = data[mapper[model.__class__.__name__]]
    else:
        X = data[num_model_features]
    result = pd.DataFrame(index=X.index, data={'Survived': model.predict(X)})
    result.index.names = ['PassengerId']
    return result

result_folder = './predictions'
#if path.exists(result_folder):
#    os.remove(result_folder)
#os.mkdir(result_folder)

for model in models:
    y_test = apply_model(model, test, col_mapper)
    y_test.to_csv(path.join(result_folder, f'{model.__class__.__name__}_prediction.csv'))
#print(f'\n{model.__class__.__name__}:')

models = [
    LogisticRegression(penalty='l1', solver='saga', random_state=0),\
    SVC(kernel='rbf'),\
    KNeighborsClassifier(),\
    GaussianNB(), MultinomialNB(), ComplementNB(), BernoulliNB(),\
    RandomForestClassifier(random_state=0), ExtraTreesClassifier(random_state=0), DecisionTreeClassifier(random_state=0),\
    XGBClassifier(random_state=0, verbosity=0), 
    LGBMClassifier(silent=True, verbose=-100), 
    CatBoostClassifier(random_state=0, silent=True)
]
for model in models:
    acc_score = check_model_100(model, data, col_mapper)
    print(f'\n{model.__class__.__name__}:')
    print(f'mean_score = {np.mean(acc_score)}; acc_score = {acc_score}')

result_folder = './predictions'
#if path.exists(result_folder):
#    os.remove(result_folder)
#os.mkdir(result_folder)

for model in models:
    y_test = apply_model(model, test, col_mapper)
    y_test.to_csv(path.join(result_folder, f'{model.__class__.__name__}_prediction_100.csv'))
#print(f'\n{model.__class__.__name__}:')
