==> 4.1.StolenCheat.ipynb <==
It's a copy of these notebook: https://www.kaggle.com/tarunpaparaju/titanic-competition-how-top-lb-got-their-score/notebook

import numpy as np
import pandas as pd

import os
import re
import warnings

cheat_dir = './special_data'
source_dir = './source_data'

test_data_with_labels = pd.read_csv(os.path.join(cheat_dir, 'cheat.csv'))
test_data = pd.read_csv(os.path.join(source_dir, 'test.csv'))

test_data_with_labels.head()

test_data.head()

warnings.filterwarnings('ignore')

for i, name in enumerate(test_data_with_labels['name']):
    if '"' in name:
        test_data_with_labels['name'][i] = re.sub('"', '', name)
        
for i, name in enumerate(test_data['Name']):
    if '"' in name:
        test_data['Name'][i] = re.sub('"', '', name)

survived = []

for name in test_data['Name']:
    survived.append(int(test_data_with_labels.loc[test_data_with_labels['name'] == name]['survived'].values[-1]))

result_dir = './predictions'
submission = pd.read_csv(os.path.join(source_dir, 'gender_submission.csv'))
submission['Survived'] = survived
submission.to_csv(os.path.join(result_dir, 'final_submission.csv'), index=False)

submission
