==> 5.3.GetSurvivedNamesFromSite.ipynb <==
# Get names of survived people

import numpy as np
import pandas as pd
from os import path
import regex
import re

Now let's try to match the names of survived people (from https://www.encyclopedia-titanica.org/titanic-survivors/) and the names from data set. First let's get the names of survived people.

survived_data_dir = './special_data'
raw_survived = ''
with open(path.join(survived_data_dir, 'encyclopedia-titanica.html'), 'r', encoding='UTF-8') as f:
    raw_survived = f.read()

last_names = regex.findall('<span itemprop="familyName">.*\,', raw_survived)
last_names = [str.replace(last_name, '<span itemprop="familyName">', '') for last_name in last_names]
last_names = [str.replace(last_name, '</span>,', '').capitalize() for last_name in last_names]
first_names = regex.findall('<span itemprop="givenName">.*<\/span>', raw_survived)
first_names = [str.replace(first_name, '<span itemprop="givenName">', '') for first_name in first_names]
first_names = [str.replace(first_name, '</span>', '') for first_name in first_names]
titles = regex.findall('<span itemprop="honorificPrefix">.*<\/span>', raw_survived)
titles = [str.replace(title, '<span itemprop="honorificPrefix">', '') for title in titles]
titles = [str.replace(title, '</span>', '') for title in titles]

titles_firsts = [' '.join(title_first) for title_first in zip(titles, first_names)]
names = [', '.join(first_last) for first_last in zip(last_names, titles_firsts)]
names

Now let's get the persons from the task

data_folder = './source_data'
raw_train = pd.read_csv(path.join(data_folder, 'train.csv'), index_col='PassengerId').rename(str.lower, axis='columns')
raw_test = pd.read_csv(path.join(data_folder, 'test.csv'), index_col='PassengerId').rename(str.lower, axis='columns')
raw_train.index.names = raw_test.index.names = ['passengerid']
train_test = pd.concat([raw_train, raw_test])
train_test = train_test[['pclass', 'name', 'sex', 'age', 'sibsp', 'parch', 'ticket', 'fare', 'cabin', 'embarked']]
train_test = train_test.sort_values(by='name')
train_test.head(40)

Now we can see that the names are different... No way. It looks like some other cheaters matched the names from site and names from task manually. No way.

Now we remember that my best score = 0.99760 despite the cheat. Also we know that there are 712 survived people totally according to info from https://www.encyclopedia-titanica.org/titanic-survivors/. Let's have a look to my cheat. How many people survived there?

cheat = pd.read_csv(path.join('./special_data', 'cheat.csv')).rename(str.lower, axis='columns')
cheat = cheat.sort_values(by='name')
cheat['survived'].sum()

Wow! Answer is much less then 712, however the accuracy is high...
